backend for statistic updated
