+++ core-service/src/main/java/ru/ssemenov/controllers/impl/CustomsDeclarationControllerImpl.java
@@ -161,9 +161,4 @@ public class CustomsDeclarationControllerImpl implements CustomsDeclarationContr
                 .headers(headers)
                 .body(resource);
     }
-
-    @GetMapping("/statistics")
-    public StatisticsDto getStatistics() {
-        return customsDeclarationServices.getStatistics();
-    }
 }
