+++ core-service/src/main/java/ru/ssemenov/controllers/impl/StatisticsControllerImpl.java
@@ -1,0 +1,11 @@
+package ru.ssemenov.controllers.impl;
+
+import ru.ssemenov.controllers.StatisticsController;
+import ru.ssemenov.dtos.StatisticsDto;
+
+public class StatisticsControllerImpl implements StatisticsController {
+    @Override
+    public StatisticsDto getStatistics() {
+        return null;
+    }
+}
